requirements.txt
#requirements is on Python version 3.11
boto3==1.36.16
botocore==1.36.16
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.18
langchain-aws==0.2.12
langchain-community==0.3.17
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langsmith==0.3.7
Markdown==3.7
#psycopg2
psycopg2-binary==2.9.10
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-slugify==8.0.4
python-socketio==5.12.1
SQLAlchemy==2.0.38
streamlit==1.42.0
urllib3==2.3.0
tabulate==0.9.0